Move bipy.core.distance exceptions to exception.py
Fixes #123.
